+++ Classes/Service/Validation/ValidatorService.php
@@ -15,12 +15,12 @@ use T3v\T3vCore\Service\AbstractService;
  */
 class ValidatorService extends AbstractService {
   /**
-   * Get the validator (factory).
+   * Gets validator (factory).
    *
    * @param string $locale The optional locale, defaults to `en_US`
    * @return \Illuminate\Validation\Factory The validator factory
    */
-  public static function getValidator(string $locale = 'en_US') {
+  public static function getValidator(string $locale = 'en_US'): Factory {
     $translator       = new Translator($locale, new MessageSelector());
     $validatorFactory = new Factory($translator);